% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_em.R
\name{calc_new_beta}
\alias{calc_new_beta}
\title{Calculates updated beta given W, X, Y}
\usage{
calc_new_beta(W, X, Y)
}
\arguments{
\item{W}{weights}

\item{X}{X matrix}

\item{Y}{Y matrix}
}
\value{
updated beta
}
\description{
Calculates updated beta given W, X, Y
}
