% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_em.R
\name{calc_new_sigma}
\alias{calc_new_sigma}
\title{Calculates updated sigma given W, X, Y, Beta}
\usage{
calc_new_sigma(W, X, Y, Beta)
}
\arguments{
\item{W}{weights}

\item{X}{X matrix}

\item{Y}{Y matrix}

\item{Beta}{updated betas}
}
\value{
updated sigma
}
\description{
Calculates updated sigma given W, X, Y, Beta
}
