% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_trajectory_data.R
\name{cluster_trajectory_data}
\alias{cluster_trajectory_data}
\title{Clusters trajectory data}
\usage{
cluster_trajectory_data(trajectory_data, P = 3, K = 5, niter = 20)
}
\arguments{
\item{trajectory_data}{data to cluster}

\item{P}{degree of bezier curve}

\item{K}{number of clusters}

\item{niter}{number of iteration to run it for}
}
\value{
results of the clustering
}
\description{
Clusters trajectory data
}
