% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_plays.R
\name{cut_plays}
\alias{cut_plays}
\title{cut plays to only get the route information we care about}
\usage{
cut_plays(data)
}
\arguments{
\item{data}{routes from the tracking data}
}
\value{
the routes cut up at the appropriate places
}
\description{
cut plays to only get the route information we care about
}
