% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_em_nested.R
\name{driver_em_nested}
\alias{driver_em_nested}
\title{Analyse nested trajectory data}
\usage{
driver_em_nested(nest_traj, K)
}
\arguments{
\item{nest_traj}{Tibble containing the nested trajectory data with at least columns x, y}

\item{K}{Number of clusters to identify via the Expectation Maximization Algorithm}
}
\value{
Tibble containing results of the EM algorithm
}
\description{
Analyse nested trajectory data
}
