% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cluster_means.R
\name{extract_cluster_means}
\alias{extract_cluster_means}
\title{extracts the cluster means from em_results}
\usage{
extract_cluster_means(em_results)
}
\arguments{
\item{em_results}{results of the em algorithm}
}
\value{
the extract clustered means
}
\description{
extracts the cluster means from em_results
}
