% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_nfl_data.R
\name{format_nfl_data}
\alias{format_nfl_data}
\title{Read and format data from csv for NFL examples, assuming BDB or 903124 format (NextGenStats)}
\usage{
format_nfl_data(file_name, data_source = "ngs")
}
\arguments{
\item{file_name}{Take a file name as it is required input to the csv parser}

\item{data_source}{either "ngs" or "903124" can we figure out how to make these the only two options?}
}
\value{
A tibble containing the original routes data formatted for use with functions in this package
}
\description{
Read and format data from csv for NFL examples, assuming BDB or 903124 format (NextGenStats)
}
