% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_em.R
\name{generate_random_cluster_controls}
\alias{generate_random_cluster_controls}
\title{Generate random controls for different template curves}
\usage{
generate_random_cluster_controls(number_of_clusters = 5, P = 3)
}
\arguments{
\item{number_of_clusters}{number of template curves}

\item{P}{Degree of the bezier curve}
}
\value{
The randomized cluster controls
}
\description{
Generate random controls for different template curves
}
