% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_data.R
\name{generate_sample_data}
\alias{generate_sample_data}
\title{Generate sample data from the  template curves}
\usage{
generate_sample_data(cluster_controls, P = 3, number_of_curves = 120)
}
\arguments{
\item{cluster_controls}{cluster controls for the random curves}

\item{P}{Degree of the bezier curve}

\item{number_of_curves}{the number of curves you want to generate}
}
\value{
The data describing the randomized curves from the clusters
}
\description{
Generate sample data from the  template curves
}
