% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_clusters.R
\name{identify_clusters}
\alias{identify_clusters}
\title{Identify the curves assigned to each cluster via Expectation Maximization Algorithm}
\usage{
identify_clusters(nest_traj, em_results)
}
\arguments{
\item{nest_traj}{Tibble containing the nested trajectory data with at least columns x, y}

\item{em_results}{The output of driver_em_nested()}
}
\value{
tibble containing the original input and the assigned clusters from the EM algorithm
}
\description{
Identify the curves assigned to each cluster via Expectation Maximization Algorithm
}
