% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_new_data.R
\name{new_data_probs}
\alias{new_data_probs}
\title{calculate the probability of assignment for new data}
\usage{
new_data_probs(routes_data, Alpha, Beta, Sigma)
}
\arguments{
\item{routes_data}{the new trajectory data}

\item{Alpha}{The alpha from the emclustering}

\item{Beta}{the beta from the em clustering}

\item{Sigma}{the sigma from the em clustering}
}
\value{
probabilities for each new cluster
}
\description{
calculate the probability of assignment for new data
}
