% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_means.R
\name{plot_cluster_means}
\alias{plot_cluster_means}
\title{plot cluster means given the results of the em}
\usage{
plot_cluster_means(em_results)
}
\arguments{
\item{em_results}{results of the em algorithm}
}
\value{
plot of the cluster means
}
\description{
plot cluster means given the results of the em
}
