% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curve_assign.R
\name{plot_curve_assign}
\alias{plot_curve_assign}
\title{Generate a facet wrapped ggplot2 object of the curves and their matched cluster mean}
\usage{
plot_curve_assign(nest_traj_with_em)
}
\arguments{
\item{nest_traj_with_em}{The output of identify_clusters, contains the raw tracking data and curve assignments}
}
\value{
ggplot2 object of the curves and their matched cluster mean
}
\description{
Generate a facet wrapped ggplot2 object of the curves and their matched cluster mean
}
