% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_centred_curves.R
\name{rand_centred_curves}
\alias{rand_centred_curves}
\title{Create a number of curves from a provided number of clusters}
\usage{
rand_centred_curves(n_clust, n_curves)
}
\arguments{
\item{n_clust}{the number of clusters to define randomly}

\item{n_curves}{the number of curves to define about the clusters randomly}
}
\value{
Tibble containing randomly generated data about the provided number of cluster means
}
\description{
Create a number of curves from a provided number of clusters
}
