% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_em.R
\name{reorder_clusters}
\alias{reorder_clusters}
\title{reorder clusters so that the biggest ones are numbered 1 and smallest n}
\usage{
reorder_clusters(em_results)
}
\arguments{
\item{em_results}{results of em}
}
\value{
reordered em_results
}
\description{
reorder clusters so that the biggest ones are numbered 1 and smallest n
}
