% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/nfl_data_functions.R
=======
% Please edit documentation in R/cut_plays.R
>>>>>>> 6ec88e675312e585c05fccf8b2fabc96d07d6261
\name{cut_plays}
\alias{cut_plays}
\title{cut plays to only get the route information we care about}
\usage{
cut_plays(data)
}
\arguments{
\item{data}{routes from the tracking data}
}
\value{
the routes cut up at the appropriate places
}
\description{
cut plays to only get the route information we care about
}
