% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_new_data.R
\name{fit_new_data}
\alias{fit_new_data}
\title{Assign clusters to never before seen data}
\usage{
fit_new_data(new_trajectory_data, em_results)
}
\arguments{
\item{new_trajectory_data}{the new data to cluster}

\item{em_results}{the results of the previous clustering}
}
\value{
<<<<<<< HEAD
The clusters for the never before seen data
=======
The clusters for the new data
>>>>>>> 6ec88e675312e585c05fccf8b2fabc96d07d6261
}
\description{
Assign clusters to never before seen data
}
