% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/nfl_data_functions.R
=======
% Please edit documentation in R/flip_field.R
>>>>>>> 6ec88e675312e585c05fccf8b2fabc96d07d6261
\name{flip_field}
\alias{flip_field}
\title{flip field about the split line}
\usage{
flip_field(data, team, direction_left, line_of_scrimmage)
}
\arguments{
\item{data}{routes from the tracking data}

\item{team}{the team the player plays on}

\item{direction_left}{the team that is going in the left direction}

\item{line_of_scrimmage}{what the line of scrimmage is for tat play}
}
\value{
return te data that has been flipped about the split line
}
\description{
flip field about the split line
}
