% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/nfl_data_functions.R
=======
% Please edit documentation in R/flip_field_903124.R
>>>>>>> 6ec88e675312e585c05fccf8b2fabc96d07d6261
\name{flip_field_903124}
\alias{flip_field_903124}
\title{flip field about the split line}
\usage{
<<<<<<< HEAD
flip_field_903124(data, left, line_of_scrimmage)
=======
flip_field_903124(data, team, left, line_of_scrimmage)
>>>>>>> 6ec88e675312e585c05fccf8b2fabc96d07d6261
}
\arguments{
\item{data}{routes from the tracking data}

<<<<<<< HEAD
=======
\item{team}{placeholder for now, need to adjust functionality with that of flip_field.R}

>>>>>>> 6ec88e675312e585c05fccf8b2fabc96d07d6261
\item{left}{the team the player plays on}

\item{line_of_scrimmage}{what the line of scrimmage is for tat play}
}
\value{
return te data that has been flipped about the split line
}
\description{
flip field about the split line
}
