% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_data_functions.R
\name{flip_field_903124}
\alias{flip_field_903124}
\title{flip field about the split line}
\usage{
flip_field_903124(data, left, line_of_scrimmage)
}
\arguments{
\item{data}{routes from the tracking data}

\item{left}{the team the player plays on}

\item{line_of_scrimmage}{what the line of scrimmage is for tat play}
}
\value{
return te data that has been flipped about the split line
}
\description{
flip field about the split line
}
