% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/nfl_data_functions.R
=======
% Please edit documentation in R/read_routes_from_903124.R
>>>>>>> 6ec88e675312e585c05fccf8b2fabc96d07d6261
\name{read_routes_from_903124}
\alias{read_routes_from_903124}
\title{read in route data from tracking data csv}
\usage{
read_routes_from_903124(file_name)
}
\arguments{
\item{file_name}{name of the tracking data file}
}
\value{
The routes from the route runner from the tracking file
}
\description{
read in route data from tracking data csv
}
